+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.4.8] 2020-03-26
 ### Added
-- Optional Environment variable AMQP_PREFETCH_LIMIT
+- Optional Environment variable AMQP_PREFETCH_COUNT
 
 
 ## [0.4.0] - 2020-03-04